backend/requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
simplejson==3.19.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.2.3
